requirements.txt
streamlit==1.26.0
langchain==0.0.237
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.33.0
torch==2.0.1
pydantic==1.10.12
